Fix pylint cyclic-import
