+++ podman/api/__init__.py
@@ -1,9 +1,8 @@
 """Tools for connecting to a Podman service."""
 
-import re
-
 from podman.api.cached_property import cached_property
 from podman.api.client import APIClient
+from podman.api.api_versions import VERSION, COMPATIBLE_VERSION
 from podman.api.http_utils import prepare_body, prepare_filters
 from podman.api.parse_utils import (
     decode_header,
@@ -15,23 +14,8 @@ from podman.api.parse_utils import (
     stream_helper,
 )
 from podman.api.tar_utils import create_tar, prepare_containerfile, prepare_containerignore
-from .. import version
 
 DEFAULT_CHUNK_SIZE = 2 * 1024 * 1024
-
-
-def _api_version(release: str, significant: int = 3) -> str:
-    """Return API version removing any additional identifiers from the release version.
-
-    This is a simple lexicographical parsing, no semantics are applied, e.g. semver checking.
-    """
-    items = re.split(r"\.|-|\+", release)
-    parts = items[0:significant]
-    return ".".join(parts)
-
-
-VERSION: str = _api_version(version.__version__)
-COMPATIBLE_VERSION: str = _api_version(version.__compatible_version__, 2)
 
 try:
     from typing import Literal
